+++ CHANGELOG.md
@@ -13,6 +13,9 @@
   favour of `StdError::generic_err` and friends.
 - Implement `From<&[u8; $N]> for Binary` and `From<[u8; $N]> for Binary` for all
   `$N <= 32`.
+- Add `Context` object that can be used to build Init/Handle/Migrate response
+  via `add_log`, `add_message`, `set_data` and then convert to the proper type
+  via `into` or `try_into`. Option to simplify response construction.
 
 **cosmwasm-vm**